+++ CHANGES.md
@@ -11,10 +11,12 @@
 
 ## develop
 
-- [ADD] OnDataChannel を追加
-    - @melpon
 - [UPDATE] Sora C++ SDK を `2022.9.0` に上げる
     - @melpon
+- [ADD] OnDataChannel を追加
+    - @melpon
+- [FIX] Sora Unity SDK のクライアント情報を設定するように追加
+    - @torikizi
 - [FIX] UnityAudioOutput = true で ADM を Stop しても再生中のままになっていたのを修正
     - @melpon
 - [FIX] ProcessAudio が機能してなかったのを修正